garmin-sync/requirements.txt
garminconnect>=0.3.2,<0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0
